first interface

+++ app/src/main/java/com/samuvlad/tempo/data/api/model/weather/Coord.kt
@@ -1,3 +1,3 @@
 package com.samuvlad.tempo.data.api.model.weather
 
-data class Coord(val lon: Long, val lat: Long)
+data class Coord(val lon: Double, val lat: Double)
